requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
openai>=1.0.0

# CRM Integration Dependencies (Optional)
# Uncomment the ones you need based on your CRM choice:
# hubspot-api-client>=7.0.0  # For HubSpot integration
# simple-salesforce>=1.12.0  # For Salesforce integration
